cleanup: move packages to top level, change vanity URL
