+++ .golangci.yml
@@ -1,7 +1,7 @@
 run:
   skip-dirs:
     # progressui is a modified 3rd party library from buildkit
-    - pkg/progressui
+    - util/progressui
 
 linters:
   disable-all: true
