+++ cmd/dagger/cmd/input/dir.go
@@ -4,11 +4,11 @@ import (
 	"path/filepath"
 	"strings"
 
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
-	"dagger.io/go/dagger/state"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
+	"go.dagger.io/dagger/state"
 )
 
 var dirCmd = &cobra.Command{
