+++ cmd/dagger/cmd/input/list.go
@@ -6,13 +6,13 @@ import (
 	"os"
 	"text/tabwriter"
 
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
-	"dagger.io/go/dagger/client"
-	"dagger.io/go/dagger/compiler"
-	"dagger.io/go/dagger/environment"
-	"dagger.io/go/dagger/solver"
-	"dagger.io/go/dagger/state"
+	"go.dagger.io/dagger/client"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
+	"go.dagger.io/dagger/compiler"
+	"go.dagger.io/dagger/environment"
+	"go.dagger.io/dagger/solver"
+	"go.dagger.io/dagger/state"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
