+++ cmd/dagger/cmd/input/secret.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"syscall"
 
-	"dagger.io/go/cmd/dagger/logger"
-	"dagger.io/go/dagger/state"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"go.dagger.io/dagger/cmd/dagger/logger"
+	"go.dagger.io/dagger/state"
 	"golang.org/x/term"
 )
