+++ cmd/dagger/cmd/list.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
 )
 
 var listCmd = &cobra.Command{
