+++ cmd/dagger/cmd/new.go
@@ -1,10 +1,10 @@
 package cmd
 
 import (
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
 )
 
 var newCmd = &cobra.Command{
