+++ cmd/dagger/cmd/query.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 
 	"cuelang.org/go/cue"
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
-	"dagger.io/go/dagger/client"
-	"dagger.io/go/dagger/compiler"
+	"go.dagger.io/dagger/client"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
+	"go.dagger.io/dagger/compiler"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
