+++ cmd/dagger/cmd/up.go
@@ -1,8 +1,8 @@
 package cmd
 
 import (
-	"dagger.io/go/cmd/dagger/cmd/common"
-	"dagger.io/go/cmd/dagger/logger"
+	"go.dagger.io/dagger/cmd/dagger/cmd/common"
+	"go.dagger.io/dagger/cmd/dagger/logger"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
