+++ keychain/encrypt.go
