+++ solver/fs.go
