+++ solver/registryauth.go
