+++ solver/solver.go
